requirements.txt
python-dotenv
cohere
umap-learn
altair
matplotlib
datashader
bokeh
holoviews
scikit-image
colorcet